+++ __tests__/cancel-token.ts
@@ -1,4 +1,4 @@
-import { defer, getCancelToken } from "../src/utils";
+import { defer, getCancelToken } from "../src/utils/general";
 import { waitFor } from "@testing-library/vue";
 import { AbortSignalWithPromise } from "../src/types/index";
