+++ __tests__/wait-for.ts
@@ -1,4 +1,4 @@
-import { waitForValue, timeout } from "../src/utils";
+import { waitForValue, timeout } from "../src/utils/general";
 import { mockSetup } from "../test-utils/components";
 import { ref } from "@vue/composition-api";
 import { wait } from "./task-cancel";
