+++ src/index.ts
@@ -1,6 +1,6 @@
 export { usePipeTask, useParallelTask, useSequentialTask } from "./wrap-utils";
-export { printTask, waitForValue, getCancelToken, timeout, useAsyncTask } from "./utils";
-export { useTaskPrefetch, useSSRPersistance } from "./ssr-utils";
+export { printTask, waitForValue, getCancelToken, timeout, useAsyncTask } from "./utils/general";
+export { useTaskPrefetch, useSSRPersistance } from "./utils/ssr-utils";
 
 export { default as useTaskGroup } from "./TaskGroup";
 export { default as useTask } from "./Task";
