+++ src/utils/ssr-utils.ts
@@ -3,8 +3,8 @@ import {
   getCurrentInstance,
   computed,
 } from "@vue/composition-api";
-import { TaskInstance } from "./TaskInstance";
-import { Task } from "./Task";
+import { TaskInstance } from "../TaskInstance";
+import { Task } from "../Task";
 
 const isServer = () => typeof window === "undefined";
